cleaning up the data fetching
